library: tbf: add structure
