+++ libraries/tock-tbf/src/lib.rs
@@ -4,4 +4,5 @@
 #![forbid(unsafe_code)]
 #![no_std]
 
-pub mod tbfheader;
+pub mod parse;
+pub mod types;
